requirements.txt
aiogram==2.25.2
aiohttp==3.8.6
aiosignal==1.3.2
async-timeout==4.0.3
attrs==25.3.0
Babel==2.9.1
certifi==2025.6.15
charset-normalizer==3.4.2
frozenlist==1.7.0
idna==3.10
magic-filter==1.0.12
multidict==6.5.0
propcache==0.3.2
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.4
urllib3==2.5.0
yarl==1.20.1
